refactor: change paymentDetail directories
